master 1.0

+++ src/main/java/com/xhu/onlinecourse/entity/Adm.java
@@ -11,8 +11,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Adm {
-    private int admId;
+    private Long admId;
     private String admName;
     private String admPassword;
-    private int admAuthoritylevels;
+    private Integer admAuthoritylevels;
 }
